refactoring forms
